Move verification of security council inside circuit
(cherry picked from commit b489ee0453eff0d255bd2cf8e8bfb7bcc0bd6fd2)
