+++ crates/light-client-prover/Cargo.toml
@@ -26,7 +26,7 @@ sov-rollup-interface = { path = "../sovereign-sdk/rollup-interface" }
 sov-state = { path = "../sovereign-sdk/module-system/sov-state", optional = true, features = ["native"] }
 
 # 3rd-party deps
-alloy-primitives = { workspace = true, optional = true }
+alloy-primitives = { workspace = true }
 anyhow = { workspace = true, optional = true }
 async-trait = { workspace = true, optional = true }
 borsh = { workspace = true }
@@ -43,7 +43,6 @@ tokio = { workspace = true, optional = true }
 tracing = { workspace = true, optional = true }
 
 [dev-dependencies]
-alloy-primitives = { workspace = true }
 alloy-signer = { workspace = true }
 alloy-signer-local = { workspace = true }
 rand = { workspace = true }
@@ -57,7 +56,6 @@ tempfile = { workspace = true }
 [features]
 default = []
 native = [
-  "dep:alloy-primitives",
   "dep:sov-mock-da",
   "dep:bitcoin-da",
   "dep:prover-services",
