+++ crates/sovereign-sdk/rollup-interface/Cargo.toml
@@ -15,7 +15,7 @@ readme = "README.md"
 resolver = "2"
 
 [dependencies]
-alloy-primitives = { workspace = true }
+alloy-primitives = { workspace = true, features = ["serde"], optional = true }
 anyhow = { workspace = true, features = ["default"] }
 async-trait = { workspace = true, optional = true }
 borsh = { workspace = true, features = ["default", "bytes", "rc"] }
@@ -45,7 +45,7 @@ serde_json = { workspace = true }
 [features]
 default = []
 native = [
-  "alloy-primitives/serde",
+  "alloy-primitives",
   "tokio",
   "futures",
   "tracing",
